refactor(command_manager): Simplify command_manager code (#1875)
- Simplify the registration of boolean remote command, now it can be registered
  by `ommand_manager::register_bool_command(bool &value, const std::string &command, const std::string &help)`
- Change the type of `node_live_percentage_threshold_for_update` from
  uint64 to int32, there is no impact bacause it's a percentage which
  is range in [0, 100].
- Change the type of `replica_assign_delay_ms_for_dropouts` from uint64
  to int64, there is less impact bacause int64 is big enough to set a
  milliseconds time interval (maximum to 106751991167 days).
- Change the type of `max_copy_rate_megabytes_per_disk`,
  `max_send_rate_megabytes_per_disk` to int64, there is no impact
  bacause its range is enlarged.
- Use FLAGS_max_concurrent_bulk_load_downloading_count directly and remove
  it from replication_options structure and replica_stub class.

+++ src/meta/test/meta_test_base.cpp
@@ -116,7 +116,7 @@ void meta_test_base::set_min_live_node_count_for_unfreeze(uint64_t node_count)
     FLAGS_min_live_node_count_for_unfreeze = node_count;
 }
 
-void meta_test_base::set_node_live_percentage_threshold_for_update(uint64_t percentage_threshold)
+void meta_test_base::set_node_live_percentage_threshold_for_update(int32_t percentage_threshold)
 {
     _ms->_node_live_percentage_threshold_for_update = percentage_threshold;
 }
